build: Set Codecov to informational mode again
We have informational mode enabled for pull requests. This commit
enables it for individual commits, too.

Informational mode registers the result of a Codecov check without ever
failing due to low coverage. This is appropriate for Terraform because
much of the test coverage comes from cross-package tests, which Codecov
misses.